restructure core app root
